Replace TPL absolute paths with library names
This should help with relocatability: instead of generating
eccodes-config.cmake with an absolute path to Third Party Libraries
(TPLs) (e.g. /lib/x86_64-linux-gnu/libm.so.6) we use the library
name instead (e.g. -lm).

We also use library names in eccodes-targets.cmake and eccodes.pc.

The motivation is so that we can distribute packages with conda-forge.
Without this change, the conda-forge build leaves its build path in the
files above. When we come to build another package downstream, e.g.
Magics, we are unable to find the ecCodes libraries because the absolute
paths to the TPLs are no longer valid. Using library names instead lets
cmake find the TPLs on the downstream build system.

We currently use -lpthreads instead of an absolute path, so hopefully
doing the same for the other TPLs isn't too controversial.